Fixing warning.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@652 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/GUI/EditRobotDialog.cpp
@@ -741,7 +741,7 @@ EditRobotDialog::DeleteBase() {
         for(MIT mit = rit->second.begin(); mit != rit->second.end(); ++mit)
           (*mit)->ChangeIndex(-(m_jointList->count()+1));
     }
-    for(size_t i = 0; i < m_jointList->count(); ++i) {
+    for(int i = 0; i < m_jointList->count(); ++i) {
       m_robotBody->DeleteJoint(index, 0, bodyCount+1);
       m_robotBody->DeleteBody(bodyCount+1);
     }
